Update install.sh
Disabling tests for a software like bitcoin is a really bad idea! It can have huge consequences e.g. unclaimable funds in the future.

`make check` should be added before `make install`. It runs all tests needed to verify correct build of binaries. Does not run all unit tests so does not take long to run, but is a must for verification. You would not want your software to flip e.g. the 18th bit of your pubkey just because a muon hit the cpu when it was compiling :) `make check`  handles that!